Update mazeGenerator.py
Modified mazeGenerator.py to generate environment based on the input of number of cans given through server.py. Location of cans, cups and bin are maintained in object dictionary and the same is maintained in objects.json file. Blocked edges were added so that it will be useful in navigation.